factroy, widget, modelに分離
