+++ src/lib/link/factory.tsx
@@ -4,8 +4,9 @@
 	を定義する
 */
 
-import { DefaultLinkFactory } from "@projectstorm/react-diagrams";
-import { AdvancedLinkModel, AdvancedLinkSegment } from "./model";
+import { DefaultLinkFactory, DiagramEngine } from "@projectstorm/react-diagrams";
+import { AdvancedLinkModel} from "./model";
+import { AdvancedLinkSegment } from "./widget";
 
 
 export class AdvancedLinkFactory extends DefaultLinkFactory {
@@ -20,7 +21,7 @@ export class AdvancedLinkFactory extends DefaultLinkFactory {
 	generateLinkSegment(model: AdvancedLinkModel, selected: boolean, path: string) {
 		return (
 			<g>
-				<AdvancedLinkSegment model={model} path={path} />
+				<AdvancedLinkSegment engine={this.engine as DiagramEngine} node={model} path={path} />
 			</g>
 		);
 	}
